requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asyncio-dgram==2.2.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
discord-webhook==1.4.1
dnspython==2.7.0
frozenlist==1.7.0
idna==3.10
mcstatus==12.0.2
minecraft==0.1.2
multidict==6.6.3
mutf8==1.0.6
pillow==11.3.0
propcache==0.3.2
pycparser==2.22
PyNBT==3.1.0
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1